requirements.txt
PyQt6>=6.4.0
telethon>=1.34.0
aiohttp>=3.9.0
aiofiles>=23.0.0
requests>=2.31.0
Pillow>=10.0.0
mutagen>=1.47.0
apscheduler>=3.10.0
cryptg>=0.4.0
openai>=1.0.0
pyinstaller>=6.0.0
PySocks>=1.7.1
